+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for marshal-contt
 
+## 0.1.2.1 -- 2019-09-14
+
+* Fix documentation bugs.
+
 ## 0.1.2.0 -- 2019-09-14
 
 * Reimplement `calloc` and friends in terms of `alloca` to make them